refactor(io): add function and improve error handling
