+++ src/lib/server/prisma.ts
@@ -1,5 +1,6 @@
 import { PrismaClient } from '$lib/../generated/prisma/client';
-import { DATABASE_URL } from '$env/static/private';
+
+const DATABASE_URL = process.env.DATABASE_URL;
 
 const prisma = new PrismaClient({
 	datasourceUrl: DATABASE_URL
